Shims should use intptr_t for file descriptors.

+++ src/Common/src/Interop/Unix/System.Native/Interop.MksTemps.cs
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using System;
 using System.Runtime.InteropServices;
 
 internal static partial class Interop
@@ -8,7 +9,7 @@ internal static partial class Interop
     internal static partial class Sys
     {
         [DllImport(Libraries.SystemNative, SetLastError = true)]
-        internal static extern int MksTemps(
+        internal static extern IntPtr MksTemps(
             byte[] template, 
             int suffixlen);
     }
