+++ src/Native/System.Net.Http.Native/pal_multi.h
@@ -70,7 +70,7 @@ Returns CURLM_OK on success, otherwise an error code.
 isExtraFileDescriptorActive is set to a value indicating whether extraFileDescriptor has new data received.
 isTimeout is set to a value indicating whether a timeout was encountered before any file descriptors had events occur.
 */
-extern "C" int32_t MultiWait(CURLM* multiHandle, int32_t extraFileDescriptor, int32_t* isExtraFileDescriptorActive, int32_t* isTimeout);
+extern "C" int32_t MultiWait(CURLM* multiHandle, intptr_t extraFileDescriptor, int32_t* isExtraFileDescriptorActive, int32_t* isTimeout);
 
 /*
 Reads/writes available data from each easy handle.
